refactor(admin): mover componentes a carpetas por responsabilidad
